Change verifiers Add command execution to verifier controller

+++ converter/src/main/java/ru/touchin/converter/verifiers/LengthVerifier.kt
@@ -1,13 +1,14 @@
 package ru.touchin.converter.verifiers
 
 import ru.touchin.converter.commands.Command
+import java.math.BigDecimal
 
-class LengthVerifier(val maxLength: Int) : Verifier<Int> {
+class LengthVerifier(val maxLength: Int) : Verifier<BigDecimal> {
 
-    override fun verify(text: String): Command<Int> = if (text.length <= maxLength) {
+    override fun verify(text: String): Command<BigDecimal> = if (text.length <= maxLength) {
         Command.Success()
     } else {
-        Command.Remove(1)
+        Command.Remove(1.toBigDecimal())
     }
 
 }
